Pass non-primitive types to C as 'borrowed' where appropriate
Parameters passed to C functions are considered to be owned by the
receiving function. As such, the receiving function must decrement the
parameter's reference count to avoid memory leaks.

An alternative is to pass parameters as borrowed by using the `@&`
prefix in their type. The receiving function does not need to decrement
the parameter's reference count but it must only be passed to other
non-consuming functions.