associative alpha
This updates the alpha blending formula to match a formula found in a blog post that is associative and proves it.

This also updates batteries to v4.14.0 and adds mathlib v4.14.0 as a dependency.
This way we can import mathlib for the linarith tactic that is useful for proving that alpha blending is associcative.

We also update the screen size for image denotation so we can see the FPS