Update to Lean v4.20.1 and build using system toolchain (#53)
* Update to lean v4.20.1

* Use system toolchain to build lean native objects

Previously we used the bundled lean clang toolchain to build native
objects (e.g raylib_bindings.o, libleanffi.a, raylean executable).

This causes problems because we link with raylib.a which is build by the
system toolchain. There can be linking errors caused by the lean
toolchain and the system toolchain linking against different libraries,
e.g glibc.

This changes uses the LEAN_CC feature of lake to build lean native
objects using the system toolchain. So both the lean and the raylib
native objects are built using the same toolchain.

The system toolchain needs to be able to find the system libgmp and
libuv to link the lean executable.

These are now passed using the LIBRARY_PATH variable in `lake build`. On
macOS these can be installed using:

```
brew install gmp
brew install libuv
```

The Just file constructs the LIBRARY_PATH using `brew --prefix`.

On linux the system toolchain may already be able to find libgmp and
libuv (when they're installed using a distro package manager for
example). If not then the user must set the `library_path` variable when
running the justfile.

* Install gmp, libuv in the github build

* Add libuv1 and libc++ to ubuntu build deps

* Use clang on linux

* Remove libc++abi-dev libc++-dev linux dependencies in CI